Remove global hooks from tests and deploy contracts from within the specific tests
